+++ app/settings/profile/name/-components/name/component.ts
@@ -1,28 +1,11 @@
-import { tagName } from '@ember-decorators/component';
-import { action } from '@ember-decorators/object';
-import { alias } from '@ember-decorators/object/computed';
-import { service } from '@ember-decorators/service';
+import { classNames, tagName } from '@ember-decorators/component';
 import Component from '@ember/component';
-import RouterService from '@ember/routing/router-service';
-import I18n from 'ember-i18n/services/i18n';
-import Toast from 'ember-toastr/services/toast';
 
 import User from 'ember-osf-web/models/user';
-import Analytics from 'ember-osf-web/services/analytics';
-import CurrentUser from 'ember-osf-web/services/current-user';
 
-@tagName('') // No div
+@tagName('div')
+@classNames('panel-body')
 export default class ProfileNameSettings extends Component {
-    @service analytics!: Analytics;
-    @service currentUser!: CurrentUser;
-    @service i18n!: I18n;
-    @service router!: RouterService;
-    @service toast!: Toast;
-
-    @alias('currentUser.user') user!: User;
-
-    @action
-    onSave(user: User) {
-        return user;
-    }
+    // Required Parameter
+    user!: User;
 }